FT8 cold call: don't reply if the caller is not decoded If it's the first time we've seen this caller, and he has a complex callsign so that it comes in as a type 4 message, the hash may not get decoded: i.e. we receive "MYCALL <...> AA01". Replying with " MYCALL AB02" would be useless: don't do that.
It's best if the caller has been seen calling CQ or engaging in
other QSOs first: then our hashtable will be populated, and the type 4
call to us will be fully decoded.

But it may also be that this is a bug on the caller's side: it
should be possible to send a call to us that can be fully decoded.